mux/telnet: Handle telnet specific stuff in telnet code
The functions are moved mostly unmodified to telnet.c. Parsing of
out-of-bound data now happens in telnet's read callback instead of
in the core.
